updated & corrected issues with Add customer
