+++ Views/Customers/Index.cshtml
@@ -39,7 +39,7 @@
 @section Scripts {
     <script src="~/lib/ag-grid-community/dist/ag-grid-community.min.js"></script>
     <script src="~/js/commons.js"></script>
-    <script src="~/js/customer.js"></script>
+    <script src="~/js/customerlist.js"></script>
     <script>
         // grid options
         const gridOptions = {
