+++ Views/Shared/_Layout.cshtml
@@ -119,7 +119,7 @@
                 <form action="#" class="d-none d-sm-inline-block">
                 </form>
                 <div class="app-title">
-                    Export & Domestic Application
+                    Export Application
                 </div>
                 <div class="navbar-collapse collapse">
                     <ul class="navbar-nav ms-auto">
